Use RelevantInformationExtractionChain to replace RefineDocumentChain
